+++ src/Kambaz/Courses/Piazza/ViewPost/FollowupDiscussions/FollowupDiscussions.tsx
@@ -1,10 +1,8 @@
 import "./FollowupDiscussions.css";
-import { useState } from "react";
-import { createDiscussion } from "../../services/followupDiscussionService";
-import { FollowupDiscussion as FollowupDiscussionType, Post } from "../../../../types";
+import { Post } from "../../../../types";
 import FollowupDiscussion from "./FollowupDiscussion";
-import { addDiscussionToPost } from "../../services/postService";
 import EditorComponent from "../EditorComponent";
+import useFollowupDiscussions from "../../hooks/useFollowupDiscussions";
 
 interface FollowupDiscussionsProps {
   convoExists: boolean;
@@ -16,45 +14,15 @@ interface FollowupDiscussionsProps {
 // Component for displaying followup discussions of a post.
 export default function FollowupDiscussions(props: FollowupDiscussionsProps) {
 
-  const { convoExists, fudIds, postId, setPost } = props;
+  const { convoExists, fudIds, setPost, postId } = props;
 
-  // boolean indicating if a new discussion is being created
-  const [startingNewDiscussion, setStartingNewDiscussion] = useState<boolean>(false);
-
-  // boolean for keeping track of the content of a new discussion
-  const [discussionContent, setDiscussionContent] = useState<string>("");
-
-  // function to handle creation/saving of a new discussion
-  const handleOnSave = async (newDiscussionContent: string) => {
-    try {
-      // convert HTML content from React Quill to plain text before saving in database 
-      const doc = new DOMParser().parseFromString(newDiscussionContent, "text/html");
-      const plainTextContent = doc.body.textContent || "";
-
-      const newDiscussion: FollowupDiscussionType = {
-        postId: postId,
-        authorId: "123",
-        datePosted: new Date().toDateString(),
-        content: plainTextContent.trim(),
-        replies: []
-      };
-      // create the followup discussion in the db
-      const newDiscussionFromDb = await createDiscussion(newDiscussion);
-
-      if (newDiscussionFromDb?._id) {
-        // add the new discussion to the post list of fudIds
-        const updatedPost = await addDiscussionToPost(postId, newDiscussionFromDb._id);
-        // set the post that's rendering to be the updated post
-        setPost(updatedPost);
-        // clear the discussion content for the next new discussion 
-        setDiscussionContent("");
-      }
-    } catch (error) {
-      console.error("Error creating followup discussion:", error);
-    }
-    // close the editor component
-    setStartingNewDiscussion(false);
-  }
+  const {
+    startingNewDiscussion,
+    setStartingNewDiscussion,
+    discussionContent,
+    setDiscussionContent,
+    handleOnSave
+  } = useFollowupDiscussions(setPost, postId);
 
   return (
     <article
